+++ src/components/range.ts
@@ -1,0 +1,2 @@
+export * from './range/range.component.js';
+export { default } from './range/range.component.js';